New AmeSystemSound implementation, mixer and player should working now

+++ wm/volumectrl.h
@@ -15,7 +15,7 @@ public:
 
 public slots:
 	void onSlider();
-	void onSystemVolume(qreal);
+	void onSystemVolume(int);
 
 public:
 	QSlider *slider;
